op - split preconditioning into separate file
