+++ interface/ceed-preconditioning.c
@@ -1,0 +1,1269 @@
+// Copyright (c) 2017, Lawrence Livermore National Security, LLC. Produced at
+// the Lawrence Livermore National Laboratory. LLNL-CODE-734707. All Rights
+// reserved. See files LICENSE and NOTICE for details.
+//
+// This file is part of CEED, a collection of benchmarks, miniapps, software
+// libraries and APIs for efficient high-order finite element and spectral
+// element discretizations for exascale applications. For more information and
+// source code availability see http://github.com/ceed.
+//
+// The CEED research is supported by the Exascale Computing Project 17-SC-20-SC,
+// a collaborative effort of two U.S. Department of Energy organizations (Office
+// of Science and the National Nuclear Security Administration) responsible for
+// the planning and preparation of a capable exascale ecosystem, including
+// software, applications, hardware, advanced system engineering and early
+// testbed platforms, in support of the nation's exascale computing imperative.
+
+#include <ceed/ceed.h>
+#include <ceed/backend.h>
+#include <ceed-impl.h>
+#include <math.h>
+#include <stdbool.h>
+#include <stdio.h>
+#include <string.h>
+
+/// @file
+/// Implementation of CeedOperator preconditioning interfaces
+
+/// ----------------------------------------------------------------------------
+/// CeedOperator Library Internal Preconditioning Functions
+/// ----------------------------------------------------------------------------
+/// @addtogroup CeedOperatorDeveloper
+/// @{
+
+/**
+@brief Duplicate a CeedOperator with a reference Ceed to fallback for advanced
+       CeedOperator functionality
+@param op  CeedOperator to create fallback for
+@return An error code: 0 - success, otherwise - failure
+@ref Developer
+**/
+int CeedOperatorCreateFallback(CeedOperator op) {
+  int ierr;
+
+  // Fallback Ceed
+  const char *resource, *fallback_resource;
+  ierr = CeedGetResource(op->ceed, &resource); CeedChk(ierr);
+  ierr = CeedGetOperatorFallbackResource(op->ceed, &fallback_resource);
+  CeedChk(ierr);
+  if (!strcmp(resource, fallback_resource))
+    // LCOV_EXCL_START
+    return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED,
+                     "Backend %s cannot create an operator"
+                     "fallback to resource %s", resource, fallback_resource);
+  // LCOV_EXCL_STOP
+
+  // Fallback Ceed
+  Ceed ceed_ref;
+  if (!op->ceed->op_fallback_ceed) {
+    ierr = CeedInit(fallback_resource, &ceed_ref); CeedChk(ierr);
+    ceed_ref->op_fallback_parent = op->ceed;
+    ceed_ref->Error = op->ceed->Error;
+    op->ceed->op_fallback_ceed = ceed_ref;
+  }
+  ceed_ref = op->ceed->op_fallback_ceed;
+
+  // Clone Op
+  CeedOperator op_ref;
+  ierr = CeedCalloc(1, &op_ref); CeedChk(ierr);
+  memcpy(op_ref, op, sizeof(*op_ref));
+  op_ref->data = NULL;
+  op_ref->interface_setup = false;
+  op_ref->backend_setup = false;
+  op_ref->ceed = ceed_ref;
+  ierr = ceed_ref->OperatorCreate(op_ref); CeedChk(ierr);
+  op->op_fallback = op_ref;
+
+  // Clone QF
+  CeedQFunction qf_ref;
+  ierr = CeedCalloc(1, &qf_ref); CeedChk(ierr);
+  memcpy(qf_ref, (op->qf), sizeof(*qf_ref));
+  qf_ref->data = NULL;
+  qf_ref->ceed = ceed_ref;
+  ierr = ceed_ref->QFunctionCreate(qf_ref); CeedChk(ierr);
+  op_ref->qf = qf_ref;
+  op->qf_fallback = qf_ref;
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+   @brief Build nonzero pattern for non-composite operator.
+
+   Users should generally use CeedOperatorLinearAssembleSymbolic()
+
+   @ref Developer
+**/
+int CeedSingleOperatorAssembleSymbolic(CeedOperator op, CeedInt offset,
+                                       CeedInt *rows, CeedInt *cols) {
+  int ierr;
+  Ceed ceed = op->ceed;
+  if (op->composite)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "Composite operator not supported");
+  // LCOV_EXCL_STOP
+
+  CeedElemRestriction rstr_in;
+  ierr = CeedOperatorGetActiveElemRestriction(op, &rstr_in); CeedChk(ierr);
+  CeedInt num_elem, elem_size, num_nodes, num_comp;
+  ierr = CeedElemRestrictionGetNumElements(rstr_in, &num_elem); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetElementSize(rstr_in, &elem_size); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetLVectorSize(rstr_in, &num_nodes); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetNumComponents(rstr_in, &num_comp); CeedChk(ierr);
+  CeedInt layout_er[3];
+  ierr = CeedElemRestrictionGetELayout(rstr_in, &layout_er); CeedChk(ierr);
+
+  CeedInt local_num_entries = elem_size*num_comp * elem_size*num_comp * num_elem;
+
+  // Determine elem_dof relation
+  CeedVector index_vec;
+  ierr = CeedVectorCreate(ceed, num_nodes, &index_vec); CeedChk(ierr);
+  CeedScalar *array;
+  ierr = CeedVectorGetArray(index_vec, CEED_MEM_HOST, &array); CeedChk(ierr);
+  for (CeedInt i = 0; i < num_nodes; ++i) {
+    array[i] = i;
+  }
+  ierr = CeedVectorRestoreArray(index_vec, &array); CeedChk(ierr);
+  CeedVector elem_dof;
+  ierr = CeedVectorCreate(ceed, num_elem * elem_size * num_comp, &elem_dof);
+  CeedChk(ierr);
+  ierr = CeedVectorSetValue(elem_dof, 0.0); CeedChk(ierr);
+  CeedElemRestrictionApply(rstr_in, CEED_NOTRANSPOSE, index_vec,
+                           elem_dof, CEED_REQUEST_IMMEDIATE); CeedChk(ierr);
+  const CeedScalar *elem_dof_a;
+  ierr = CeedVectorGetArrayRead(elem_dof, CEED_MEM_HOST, &elem_dof_a);
+  CeedChk(ierr);
+  ierr = CeedVectorDestroy(&index_vec); CeedChk(ierr);
+
+  // Determine i, j locations for element matrices
+  CeedInt count = 0;
+  for (int e = 0; e < num_elem; ++e) {
+    for (int comp_in = 0; comp_in < num_comp; ++comp_in) {
+      for (int comp_out = 0; comp_out < num_comp; ++comp_out) {
+        for (int i = 0; i < elem_size; ++i) {
+          for (int j = 0; j < elem_size; ++j) {
+            const CeedInt elem_dof_index_row = (i)*layout_er[0] +
+                                               (comp_out)*layout_er[1] + e*layout_er[2];
+            const CeedInt elem_dof_index_col = (j)*layout_er[0] +
+                                               (comp_in)*layout_er[1] + e*layout_er[2];
+
+            const CeedInt row = elem_dof_a[elem_dof_index_row];
+            const CeedInt col = elem_dof_a[elem_dof_index_col];
+
+            rows[offset + count] = row;
+            cols[offset + count] = col;
+            count++;
+          }
+        }
+      }
+    }
+  }
+  if (count != local_num_entries)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_MAJOR, "Error computing assembled entries");
+  // LCOV_EXCL_STOP
+  ierr = CeedVectorRestoreArrayRead(elem_dof, &elem_dof_a); CeedChk(ierr);
+  ierr = CeedVectorDestroy(&elem_dof); CeedChk(ierr);
+
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+   @brief Assemble nonzero entries for non-composite operator
+
+   Users should generally use CeedOperatorLinearAssemble()
+
+   @ref Developer
+**/
+int CeedSingleOperatorAssemble(CeedOperator op, CeedInt offset,
+                               CeedVector values) {
+  int ierr;
+  Ceed ceed = op->ceed;;
+  if (op->composite)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "Composite operator not supported");
+  // LCOV_EXCL_STOP
+
+  // Assemble QFunction
+  CeedQFunction qf;
+  ierr = CeedOperatorGetQFunction(op, &qf); CeedChk(ierr);
+  CeedInt num_input_fields, num_output_fields;
+  ierr= CeedQFunctionGetNumArgs(qf, &num_input_fields, &num_output_fields);
+  CeedChk(ierr);
+  CeedVector assembled_qf;
+  CeedElemRestriction rstr_q;
+  ierr = CeedOperatorLinearAssembleQFunction(
+           op, &assembled_qf, &rstr_q, CEED_REQUEST_IMMEDIATE); CeedChk(ierr);
+
+  CeedInt qf_length;
+  ierr = CeedVectorGetLength(assembled_qf, &qf_length); CeedChk(ierr);
+
+  CeedOperatorField *input_fields;
+  CeedOperatorField *output_fields;
+  ierr = CeedOperatorGetFields(op, &input_fields, &output_fields); CeedChk(ierr);
+
+  // Determine active input basis
+  CeedQFunctionField *qf_fields;
+  ierr = CeedQFunctionGetFields(qf, &qf_fields, NULL); CeedChk(ierr);
+  CeedInt num_eval_mode_in = 0, dim = 1;
+  CeedEvalMode *eval_mode_in = NULL;
+  CeedBasis basis_in = NULL;
+  CeedElemRestriction rstr_in = NULL;
+  for (CeedInt i=0; i<num_input_fields; i++) {
+    CeedVector vec;
+    ierr = CeedOperatorFieldGetVector(input_fields[i], &vec); CeedChk(ierr);
+    if (vec == CEED_VECTOR_ACTIVE) {
+      ierr = CeedOperatorFieldGetBasis(input_fields[i], &basis_in);
+      CeedChk(ierr);
+      ierr = CeedBasisGetDimension(basis_in, &dim); CeedChk(ierr);
+      ierr = CeedOperatorFieldGetElemRestriction(input_fields[i], &rstr_in);
+      CeedChk(ierr);
+      CeedEvalMode eval_mode;
+      ierr = CeedQFunctionFieldGetEvalMode(qf_fields[i], &eval_mode);
+      CeedChk(ierr);
+      switch (eval_mode) {
+      case CEED_EVAL_NONE:
+      case CEED_EVAL_INTERP:
+        ierr = CeedRealloc(num_eval_mode_in + 1, &eval_mode_in); CeedChk(ierr);
+        eval_mode_in[num_eval_mode_in] = eval_mode;
+        num_eval_mode_in += 1;
+        break;
+      case CEED_EVAL_GRAD:
+        ierr = CeedRealloc(num_eval_mode_in + dim, &eval_mode_in); CeedChk(ierr);
+        for (CeedInt d=0; d<dim; d++) {
+          eval_mode_in[num_eval_mode_in+d] = eval_mode;
+        }
+        num_eval_mode_in += dim;
+        break;
+      case CEED_EVAL_WEIGHT:
+      case CEED_EVAL_DIV:
+      case CEED_EVAL_CURL:
+        break; // Caught by QF Assembly
+      }
+    }
+  }
+
+  // Determine active output basis
+  ierr = CeedQFunctionGetFields(qf, NULL, &qf_fields); CeedChk(ierr);
+  CeedInt num_eval_mode_out = 0;
+  CeedEvalMode *eval_mode_out = NULL;
+  CeedBasis basis_out = NULL;
+  CeedElemRestriction rstr_out = NULL;
+  for (CeedInt i=0; i<num_output_fields; i++) {
+    CeedVector vec;
+    ierr = CeedOperatorFieldGetVector(output_fields[i], &vec); CeedChk(ierr);
+    if (vec == CEED_VECTOR_ACTIVE) {
+      ierr = CeedOperatorFieldGetBasis(output_fields[i], &basis_out);
+      CeedChk(ierr);
+      ierr = CeedOperatorFieldGetElemRestriction(output_fields[i], &rstr_out);
+      CeedChk(ierr);
+      CeedEvalMode eval_mode;
+      ierr = CeedQFunctionFieldGetEvalMode(qf_fields[i], &eval_mode);
+      CeedChk(ierr);
+      switch (eval_mode) {
+      case CEED_EVAL_NONE:
+      case CEED_EVAL_INTERP:
+        ierr = CeedRealloc(num_eval_mode_out + 1, &eval_mode_out); CeedChk(ierr);
+        eval_mode_out[num_eval_mode_out] = eval_mode;
+        num_eval_mode_out += 1;
+        break;
+      case CEED_EVAL_GRAD:
+        ierr = CeedRealloc(num_eval_mode_out + dim, &eval_mode_out); CeedChk(ierr);
+        for (CeedInt d=0; d<dim; d++) {
+          eval_mode_out[num_eval_mode_out+d] = eval_mode;
+        }
+        num_eval_mode_out += dim;
+        break;
+      case CEED_EVAL_WEIGHT:
+      case CEED_EVAL_DIV:
+      case CEED_EVAL_CURL:
+        break; // Caught by QF Assembly
+      }
+    }
+  }
+
+  if (num_eval_mode_in == 0 || num_eval_mode_out == 0)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "Cannot assemble operator with out inputs/outputs");
+  // LCOV_EXCL_STOP
+
+  CeedInt num_elem, elem_size, num_qpts, num_comp;
+  ierr = CeedElemRestrictionGetNumElements(rstr_in, &num_elem); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetElementSize(rstr_in, &elem_size); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetNumComponents(rstr_in, &num_comp); CeedChk(ierr);
+  ierr = CeedBasisGetNumQuadraturePoints(basis_in, &num_qpts); CeedChk(ierr);
+
+  CeedInt local_num_entries = elem_size*num_comp * elem_size*num_comp * num_elem;
+
+  // loop over elements and put in data structure
+  const CeedScalar *interp_in, *grad_in;
+  ierr = CeedBasisGetInterp(basis_in, &interp_in); CeedChk(ierr);
+  ierr = CeedBasisGetGrad(basis_in, &grad_in); CeedChk(ierr);
+
+  const CeedScalar *assembled_qf_array;
+  ierr = CeedVectorGetArrayRead(assembled_qf, CEED_MEM_HOST, &assembled_qf_array);
+  CeedChk(ierr);
+
+  CeedInt layout_qf[3];
+  ierr = CeedElemRestrictionGetELayout(rstr_q, &layout_qf); CeedChk(ierr);
+  ierr = CeedElemRestrictionDestroy(&rstr_q); CeedChk(ierr);
+
+  // we store B_mat_in, B_mat_out, BTD, elem_mat in row-major order
+  CeedScalar B_mat_in[(num_qpts * num_eval_mode_in) * elem_size];
+  CeedScalar B_mat_out[(num_qpts * num_eval_mode_out) * elem_size];
+  CeedScalar D_mat[num_eval_mode_out * num_eval_mode_in *
+                                     num_qpts]; // logically 3-tensor
+  CeedScalar BTD[elem_size * num_qpts*num_eval_mode_in];
+  CeedScalar elem_mat[elem_size * elem_size];
+  int count = 0;
+  CeedScalar *vals;
+  ierr = CeedVectorGetArray(values, CEED_MEM_HOST, &vals); CeedChk(ierr);
+  for (int e = 0; e < num_elem; ++e) {
+    for (int comp_in = 0; comp_in < num_comp; ++comp_in) {
+      for (int comp_out = 0; comp_out < num_comp; ++comp_out) {
+        for (int ell = 0; ell < (num_qpts * num_eval_mode_in) * elem_size; ++ell) {
+          B_mat_in[ell] = 0.0;
+        }
+        for (int ell = 0; ell < (num_qpts * num_eval_mode_out) * elem_size; ++ell) {
+          B_mat_out[ell] = 0.0;
+        }
+        // Store block-diagonal D matrix as collection of small dense blocks
+        for (int ell = 0; ell < num_eval_mode_in*num_eval_mode_out*num_qpts; ++ell) {
+          D_mat[ell] = 0.0;
+        }
+        // form element matrix itself (for each block component)
+        for (int ell = 0; ell < elem_size*elem_size; ++ell) {
+          elem_mat[ell] = 0.0;
+        }
+        for (int q = 0; q < num_qpts; ++q) {
+          for (int n = 0; n < elem_size; ++n) {
+            CeedInt d_in = -1;
+            for (int e_in = 0; e_in < num_eval_mode_in; ++e_in) {
+              const int qq = num_eval_mode_in*q;
+              if (eval_mode_in[e_in] == CEED_EVAL_INTERP) {
+                B_mat_in[(qq+e_in)*elem_size + n] += interp_in[q * elem_size + n];
+              } else if (eval_mode_in[e_in] == CEED_EVAL_GRAD) {
+                d_in += 1;
+                B_mat_in[(qq+e_in)*elem_size + n] +=
+                  grad_in[(d_in*num_qpts+q) * elem_size + n];
+              } else {
+                // LCOV_EXCL_START
+                return CeedError(ceed, CEED_ERROR_UNSUPPORTED, "Not implemented!");
+                // LCOV_EXCL_STOP
+              }
+            }
+            CeedInt d_out = -1;
+            for (int e_out = 0; e_out < num_eval_mode_out; ++e_out) {
+              const int qq = num_eval_mode_out*q;
+              if (eval_mode_out[e_out] == CEED_EVAL_INTERP) {
+                B_mat_out[(qq+e_out)*elem_size + n] += interp_in[q * elem_size + n];
+              } else if (eval_mode_out[e_out] == CEED_EVAL_GRAD) {
+                d_out += 1;
+                B_mat_out[(qq+e_out)*elem_size + n] +=
+                  grad_in[(d_out*num_qpts+q) * elem_size + n];
+              } else {
+                // LCOV_EXCL_START
+                return CeedError(ceed, CEED_ERROR_UNSUPPORTED, "Not implemented!");
+                // LCOV_EXCL_STOP
+              }
+            }
+          }
+          for (int ei = 0; ei < num_eval_mode_out; ++ei) {
+            for (int ej = 0; ej < num_eval_mode_in; ++ej) {
+              const int eval_mode_index = ((ei*num_comp+comp_in)*num_eval_mode_in+ej)*num_comp
+                                          +comp_out;
+              const int index = q*layout_qf[0] + eval_mode_index*layout_qf[1] +
+                                e*layout_qf[2];
+              D_mat[(ei*num_eval_mode_in+ej)*num_qpts + q] += assembled_qf_array[index];
+            }
+          }
+        }
+        // Compute B^T*D
+        for (int ell = 0; ell < elem_size*num_qpts*num_eval_mode_in; ++ell) {
+          BTD[ell] = 0.0;
+        }
+        for (int j = 0; j<elem_size; ++j) {
+          for (int q = 0; q<num_qpts; ++q) {
+            int qq = num_eval_mode_out*q;
+            for (int ei = 0; ei < num_eval_mode_in; ++ei) {
+              for (int ej = 0; ej < num_eval_mode_out; ++ej) {
+                BTD[j*(num_qpts*num_eval_mode_in) + (qq+ei)] +=
+                  B_mat_out[(qq+ej)*elem_size + j] * D_mat[(ei*num_eval_mode_in+ej)*num_qpts + q];
+              }
+            }
+          }
+        }
+
+        ierr = CeedMatrixMultiply(ceed, BTD, B_mat_in, elem_mat, elem_size,
+                                  elem_size, num_qpts*num_eval_mode_in); CeedChk(ierr);
+
+        // put element matrix in coordinate data structure
+        for (int i = 0; i < elem_size; ++i) {
+          for (int j = 0; j < elem_size; ++j) {
+            vals[offset + count] = elem_mat[i*elem_size + j];
+            count++;
+          }
+        }
+      }
+    }
+  }
+  if (count != local_num_entries)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_MAJOR, "Error computing entries");
+  // LCOV_EXCL_STOP
+  ierr = CeedVectorRestoreArray(values, &vals); CeedChk(ierr);
+
+  ierr = CeedVectorRestoreArrayRead(assembled_qf, &assembled_qf_array);
+  CeedChk(ierr);
+  ierr = CeedVectorDestroy(&assembled_qf); CeedChk(ierr);
+  ierr = CeedFree(&eval_mode_in); CeedChk(ierr);
+  ierr = CeedFree(&eval_mode_out); CeedChk(ierr);
+
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+   @ref Utility
+**/
+int CeedSingleOperatorAssemblyCountEntries(CeedOperator op,
+    CeedInt *num_entries) {
+  int ierr;
+  CeedElemRestriction rstr;
+  CeedInt num_elem, elem_size, num_comp;
+
+  if (op->composite)
+    // LCOV_EXCL_START
+    return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED,
+                     "Composite operator not supported");
+  // LCOV_EXCL_STOP
+  ierr = CeedOperatorGetActiveElemRestriction(op, &rstr); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetNumElements(rstr, &num_elem); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetElementSize(rstr, &elem_size); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetNumComponents(rstr, &num_comp); CeedChk(ierr);
+  *num_entries = elem_size*num_comp * elem_size*num_comp * num_elem;
+
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Common code for creating a multigrid coarse operator and level
+           transfer operators for a CeedOperator
+
+  @param[in] op_fine       Fine grid operator
+  @param[in] p_mult_fine   L-vector multiplicity in parallel gather/scatter
+  @param[in] rstr_coarse   Coarse grid restriction
+  @param[in] basis_coarse  Coarse grid active vector basis
+  @param[in] basis_c_to_f  Basis for coarse to fine interpolation
+  @param[out] op_coarse    Coarse grid operator
+  @param[out] op_prolong   Coarse to fine operator
+  @param[out] op_restrict  Fine to coarse operator
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref Developer
+**/
+static int CeedOperatorMultigridLevel_Core(CeedOperator op_fine,
+    CeedVector p_mult_fine, CeedElemRestriction rstr_coarse, CeedBasis basis_coarse,
+    CeedBasis basis_c_to_f, CeedOperator *op_coarse, CeedOperator *op_prolong,
+    CeedOperator *op_restrict) {
+  int ierr;
+  Ceed ceed;
+  ierr = CeedOperatorGetCeed(op_fine, &ceed); CeedChk(ierr);
+
+  // Check for composite operator
+  bool is_composite;
+  ierr = CeedOperatorIsComposite(op_fine, &is_composite); CeedChk(ierr);
+  if (is_composite)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
+                     "Automatic multigrid setup for composite operators not supported");
+  // LCOV_EXCL_STOP
+
+  // Coarse Grid
+  ierr = CeedOperatorCreate(ceed, op_fine->qf, op_fine->dqf, op_fine->dqfT,
+                            op_coarse); CeedChk(ierr);
+  CeedElemRestriction rstr_fine = NULL;
+  // -- Clone input fields
+  for (int i = 0; i < op_fine->qf->num_input_fields; i++) {
+    if (op_fine->input_fields[i]->vec == CEED_VECTOR_ACTIVE) {
+      rstr_fine = op_fine->input_fields[i]->elem_restr;
+      ierr = CeedOperatorSetField(*op_coarse, op_fine->input_fields[i]->field_name,
+                                  rstr_coarse, basis_coarse, CEED_VECTOR_ACTIVE);
+      CeedChk(ierr);
+    } else {
+      ierr = CeedOperatorSetField(*op_coarse, op_fine->input_fields[i]->field_name,
+                                  op_fine->input_fields[i]->elem_restr,
+                                  op_fine->input_fields[i]->basis,
+                                  op_fine->input_fields[i]->vec); CeedChk(ierr);
+    }
+  }
+  // -- Clone output fields
+  for (int i = 0; i < op_fine->qf->num_output_fields; i++) {
+    if (op_fine->output_fields[i]->vec == CEED_VECTOR_ACTIVE) {
+      ierr = CeedOperatorSetField(*op_coarse, op_fine->output_fields[i]->field_name,
+                                  rstr_coarse, basis_coarse, CEED_VECTOR_ACTIVE);
+      CeedChk(ierr);
+    } else {
+      ierr = CeedOperatorSetField(*op_coarse, op_fine->output_fields[i]->field_name,
+                                  op_fine->output_fields[i]->elem_restr,
+                                  op_fine->output_fields[i]->basis,
+                                  op_fine->output_fields[i]->vec); CeedChk(ierr);
+    }
+  }
+
+  // Multiplicity vector
+  CeedVector mult_vec, mult_e_vec;
+  ierr = CeedElemRestrictionCreateVector(rstr_fine, &mult_vec, &mult_e_vec);
+  CeedChk(ierr);
+  ierr = CeedVectorSetValue(mult_e_vec, 0.0); CeedChk(ierr);
+  ierr = CeedElemRestrictionApply(rstr_fine, CEED_NOTRANSPOSE, p_mult_fine,
+                                  mult_e_vec, CEED_REQUEST_IMMEDIATE); CeedChk(ierr);
+  ierr = CeedVectorSetValue(mult_vec, 0.0); CeedChk(ierr);
+  ierr = CeedElemRestrictionApply(rstr_fine, CEED_TRANSPOSE, mult_e_vec, mult_vec,
+                                  CEED_REQUEST_IMMEDIATE); CeedChk(ierr);
+  ierr = CeedVectorDestroy(&mult_e_vec); CeedChk(ierr);
+  ierr = CeedVectorReciprocal(mult_vec); CeedChk(ierr);
+
+  // Restriction
+  CeedInt num_comp;
+  ierr = CeedBasisGetNumComponents(basis_coarse, &num_comp); CeedChk(ierr);
+  CeedQFunction qf_restrict;
+  ierr = CeedQFunctionCreateInteriorByName(ceed, "Scale", &qf_restrict);
+  CeedChk(ierr);
+  CeedInt *num_comp_r_data;
+  ierr = CeedCalloc(1, &num_comp_r_data); CeedChk(ierr);
+  num_comp_r_data[0] = num_comp;
+  CeedQFunctionContext ctx_r;
+  ierr = CeedQFunctionContextCreate(ceed, &ctx_r); CeedChk(ierr);
+  ierr = CeedQFunctionContextSetData(ctx_r, CEED_MEM_HOST, CEED_OWN_POINTER,
+                                     sizeof(*num_comp_r_data), num_comp_r_data);
+  CeedChk(ierr);
+  ierr = CeedQFunctionSetContext(qf_restrict, ctx_r); CeedChk(ierr);
+  ierr = CeedQFunctionContextDestroy(&ctx_r); CeedChk(ierr);
+  ierr = CeedQFunctionAddInput(qf_restrict, "input", num_comp, CEED_EVAL_NONE);
+  CeedChk(ierr);
+  ierr = CeedQFunctionAddInput(qf_restrict, "scale", num_comp, CEED_EVAL_NONE);
+  CeedChk(ierr);
+  ierr = CeedQFunctionAddOutput(qf_restrict, "output", num_comp,
+                                CEED_EVAL_INTERP); CeedChk(ierr);
+
+  ierr = CeedOperatorCreate(ceed, qf_restrict, CEED_QFUNCTION_NONE,
+                            CEED_QFUNCTION_NONE, op_restrict);
+  CeedChk(ierr);
+  ierr = CeedOperatorSetField(*op_restrict, "input", rstr_fine,
+                              CEED_BASIS_COLLOCATED, CEED_VECTOR_ACTIVE);
+  CeedChk(ierr);
+  ierr = CeedOperatorSetField(*op_restrict, "scale", rstr_fine,
+                              CEED_BASIS_COLLOCATED, mult_vec);
+  CeedChk(ierr);
+  ierr = CeedOperatorSetField(*op_restrict, "output", rstr_coarse, basis_c_to_f,
+                              CEED_VECTOR_ACTIVE); CeedChk(ierr);
+
+  // Prolongation
+  CeedQFunction qf_prolong;
+  ierr = CeedQFunctionCreateInteriorByName(ceed, "Scale", &qf_prolong);
+  CeedChk(ierr);
+  CeedInt *num_comp_p_data;
+  ierr = CeedCalloc(1, &num_comp_p_data); CeedChk(ierr);
+  num_comp_p_data[0] = num_comp;
+  CeedQFunctionContext ctx_p;
+  ierr = CeedQFunctionContextCreate(ceed, &ctx_p); CeedChk(ierr);
+  ierr = CeedQFunctionContextSetData(ctx_p, CEED_MEM_HOST, CEED_OWN_POINTER,
+                                     sizeof(*num_comp_p_data), num_comp_p_data);
+  CeedChk(ierr);
+  ierr = CeedQFunctionSetContext(qf_prolong, ctx_p); CeedChk(ierr);
+  ierr = CeedQFunctionContextDestroy(&ctx_p); CeedChk(ierr);
+  ierr = CeedQFunctionAddInput(qf_prolong, "input", num_comp, CEED_EVAL_INTERP);
+  CeedChk(ierr);
+  ierr = CeedQFunctionAddInput(qf_prolong, "scale", num_comp, CEED_EVAL_NONE);
+  CeedChk(ierr);
+  ierr = CeedQFunctionAddOutput(qf_prolong, "output", num_comp, CEED_EVAL_NONE);
+  CeedChk(ierr);
+
+  ierr = CeedOperatorCreate(ceed, qf_prolong, CEED_QFUNCTION_NONE,
+                            CEED_QFUNCTION_NONE, op_prolong);
+  CeedChk(ierr);
+  ierr = CeedOperatorSetField(*op_prolong, "input", rstr_coarse, basis_c_to_f,
+                              CEED_VECTOR_ACTIVE); CeedChk(ierr);
+  ierr = CeedOperatorSetField(*op_prolong, "scale", rstr_fine,
+                              CEED_BASIS_COLLOCATED, mult_vec);
+  CeedChk(ierr);
+  ierr = CeedOperatorSetField(*op_prolong, "output", rstr_fine,
+                              CEED_BASIS_COLLOCATED, CEED_VECTOR_ACTIVE);
+  CeedChk(ierr);
+
+  // Cleanup
+  ierr = CeedVectorDestroy(&mult_vec); CeedChk(ierr);
+  ierr = CeedBasisDestroy(&basis_c_to_f); CeedChk(ierr);
+  ierr = CeedQFunctionDestroy(&qf_restrict); CeedChk(ierr);
+  ierr = CeedQFunctionDestroy(&qf_prolong); CeedChk(ierr);
+  return CEED_ERROR_SUCCESS;
+}
+
+/// ----------------------------------------------------------------------------
+/// CeedOperator Public API
+/// ----------------------------------------------------------------------------
+/// @addtogroup CeedOperatorUser
+/// @{
+
+/**
+  @brief Assemble a linear CeedQFunction associated with a CeedOperator
+
+  This returns a CeedVector containing a matrix at each quadrature point
+    providing the action of the CeedQFunction associated with the CeedOperator.
+    The vector 'assembled' is of shape
+      [num_elements, num_input_fields, num_output_fields, num_quad_points]
+    and contains column-major matrices representing the action of the
+    CeedQFunction for a corresponding quadrature point on an element. Inputs and
+    outputs are in the order provided by the user when adding CeedOperator fields.
+    For example, a CeedQFunction with inputs 'u' and 'gradu' and outputs 'gradv' and
+    'v', provided in that order, would result in an assembled QFunction that
+    consists of (1 + dim) x (dim + 1) matrices at each quadrature point acting
+    on the input [u, du_0, du_1] and producing the output [dv_0, dv_1, v].
+
+  @param op              CeedOperator to assemble CeedQFunction
+  @param[out] assembled  CeedVector to store assembled CeedQFunction at
+                           quadrature points
+  @param[out] rstr       CeedElemRestriction for CeedVector containing assembled
+                           CeedQFunction
+  @param request         Address of CeedRequest for non-blocking completion, else
+                           @ref CEED_REQUEST_IMMEDIATE
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorLinearAssembleQFunction(CeedOperator op, CeedVector *assembled,
+                                        CeedElemRestriction *rstr,
+                                        CeedRequest *request) {
+  int ierr;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Backend version
+  if (op->LinearAssembleQFunction) {
+    return op->LinearAssembleQFunction(op, assembled, rstr, request);
+  } else {
+    // Fallback to reference Ceed
+    if (!op->op_fallback) {
+      ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+    }
+    // Assemble
+    return CeedOperatorLinearAssembleQFunction(op->op_fallback, assembled,
+           rstr, request);
+  }
+}
+
+/**
+  @brief Assemble the diagonal of a square linear CeedOperator
+
+  This overwrites a CeedVector with the diagonal of a linear CeedOperator.
+
+  Note: Currently only non-composite CeedOperators with a single field and
+          composite CeedOperators with single field sub-operators are supported.
+
+  @param op              CeedOperator to assemble CeedQFunction
+  @param[out] assembled  CeedVector to store assembled CeedOperator diagonal
+  @param request         Address of CeedRequest for non-blocking completion, else
+                           @ref CEED_REQUEST_IMMEDIATE
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorLinearAssembleDiagonal(CeedOperator op, CeedVector assembled,
+                                       CeedRequest *request) {
+  int ierr;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Use backend version, if available
+  if (op->LinearAssembleDiagonal) {
+    return op->LinearAssembleDiagonal(op, assembled, request);
+  } else if (op->LinearAssembleAddDiagonal) {
+    ierr = CeedVectorSetValue(assembled, 0.0); CeedChk(ierr);
+    return CeedOperatorLinearAssembleAddDiagonal(op, assembled, request);
+  } else {
+    // Fallback to reference Ceed
+    if (!op->op_fallback) {
+      ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+    }
+    // Assemble
+    return CeedOperatorLinearAssembleDiagonal(op->op_fallback, assembled,
+           request);
+  }
+}
+
+/**
+  @brief Assemble the diagonal of a square linear CeedOperator
+
+  This sums into a CeedVector the diagonal of a linear CeedOperator.
+
+  Note: Currently only non-composite CeedOperators with a single field and
+          composite CeedOperators with single field sub-operators are supported.
+
+  @param op              CeedOperator to assemble CeedQFunction
+  @param[out] assembled  CeedVector to store assembled CeedOperator diagonal
+  @param request         Address of CeedRequest for non-blocking completion, else
+                           @ref CEED_REQUEST_IMMEDIATE
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorLinearAssembleAddDiagonal(CeedOperator op, CeedVector assembled,
+    CeedRequest *request) {
+  int ierr;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Use backend version, if available
+  if (op->LinearAssembleAddDiagonal) {
+    return op->LinearAssembleAddDiagonal(op, assembled, request);
+  } else {
+    // Fallback to reference Ceed
+    if (!op->op_fallback) {
+      ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+    }
+    // Assemble
+    return CeedOperatorLinearAssembleAddDiagonal(op->op_fallback, assembled,
+           request);
+  }
+}
+
+/**
+  @brief Assemble the point block diagonal of a square linear CeedOperator
+
+  This overwrites a CeedVector with the point block diagonal of a linear
+    CeedOperator.
+
+  Note: Currently only non-composite CeedOperators with a single field and
+          composite CeedOperators with single field sub-operators are supported.
+
+  @param op              CeedOperator to assemble CeedQFunction
+  @param[out] assembled  CeedVector to store assembled CeedOperator point block
+                           diagonal, provided in row-major form with an
+                           @a num_comp * @a num_comp block at each node. The dimensions
+                           of this vector are derived from the active vector
+                           for the CeedOperator. The array has shape
+                           [nodes, component out, component in].
+  @param request         Address of CeedRequest for non-blocking completion, else
+                           CEED_REQUEST_IMMEDIATE
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorLinearAssemblePointBlockDiagonal(CeedOperator op,
+    CeedVector assembled, CeedRequest *request) {
+  int ierr;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Use backend version, if available
+  if (op->LinearAssemblePointBlockDiagonal) {
+    return op->LinearAssemblePointBlockDiagonal(op, assembled, request);
+  } else if (op->LinearAssembleAddPointBlockDiagonal) {
+    ierr = CeedVectorSetValue(assembled, 0.0); CeedChk(ierr);
+    return CeedOperatorLinearAssembleAddPointBlockDiagonal(op, assembled,
+           request);
+  } else {
+    // Fallback to reference Ceed
+    if (!op->op_fallback) {
+      ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+    }
+    // Assemble
+    return CeedOperatorLinearAssemblePointBlockDiagonal(op->op_fallback,
+           assembled, request);
+  }
+}
+
+/**
+  @brief Assemble the point block diagonal of a square linear CeedOperator
+
+  This sums into a CeedVector with the point block diagonal of a linear
+    CeedOperator.
+
+  Note: Currently only non-composite CeedOperators with a single field and
+          composite CeedOperators with single field sub-operators are supported.
+
+  @param op              CeedOperator to assemble CeedQFunction
+  @param[out] assembled  CeedVector to store assembled CeedOperator point block
+                           diagonal, provided in row-major form with an
+                           @a num_comp * @a num_comp block at each node. The dimensions
+                           of this vector are derived from the active vector
+                           for the CeedOperator. The array has shape
+                           [nodes, component out, component in].
+  @param request         Address of CeedRequest for non-blocking completion, else
+                           CEED_REQUEST_IMMEDIATE
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorLinearAssembleAddPointBlockDiagonal(CeedOperator op,
+    CeedVector assembled, CeedRequest *request) {
+  int ierr;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Use backend version, if available
+  if (op->LinearAssembleAddPointBlockDiagonal) {
+    return op->LinearAssembleAddPointBlockDiagonal(op, assembled, request);
+  } else {
+    // Fallback to reference Ceed
+    if (!op->op_fallback) {
+      ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+    }
+    // Assemble
+    return CeedOperatorLinearAssembleAddPointBlockDiagonal(op->op_fallback,
+           assembled, request);
+  }
+}
+
+/**
+   @brief Fully assemble the nonzero pattern of a linear operator.
+
+   Expected to be used in conjunction with CeedOperatorLinearAssemble()
+
+   The assembly routines use coordinate format, with num_entries tuples of the
+   form (i, j, value) which indicate that value should be added to the matrix
+   in entry (i, j). Note that the (i, j) pairs are not unique and may repeat.
+   This function returns the number of entries and their (i, j) locations,
+   while CeedOperatorLinearAssemble() provides the values in the same
+   ordering.
+
+   This will generally be slow unless your operator is low-order.
+
+   @param[in]  op           CeedOperator to assemble
+   @param[out] num_entries  Number of entries in coordinate nonzero pattern.
+   @param[out] rows         Row number for each entry.
+   @param[out] cols         Column number for each entry.
+
+   @ref User
+**/
+int CeedOperatorLinearAssembleSymbolic(CeedOperator op,
+                                       CeedInt *num_entries, CeedInt **rows, CeedInt **cols) {
+  int ierr;
+  CeedInt num_suboperators, single_entries;
+  CeedOperator *sub_operators;
+  bool is_composite;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Use backend version, if available
+  if (op->LinearAssembleSymbolic) {
+    return op->LinearAssembleSymbolic(op, num_entries, rows, cols);
+  } else {
+    // Check for valid fallback resource
+    const char *resource, *fallback_resource;
+    ierr = CeedGetResource(op->ceed, &resource); CeedChk(ierr);
+    ierr = CeedGetOperatorFallbackResource(op->ceed, &fallback_resource);
+    CeedChk(ierr);
+    if (strcmp(fallback_resource, "") && strcmp(resource, fallback_resource)) {
+      // Fallback to reference Ceed
+      if (!op->op_fallback) {
+        ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+      }
+      // Assemble
+      return CeedOperatorLinearAssembleSymbolic(op->op_fallback, num_entries, rows,
+             cols);
+    }
+  }
+
+  // if neither backend nor fallback resource provides
+  // LinearAssembleSymbolic, continue with interface-level implementation
+
+  // count entries and allocate rows, cols arrays
+  ierr = CeedOperatorIsComposite(op, &is_composite); CeedChk(ierr);
+  *num_entries = 0;
+  if (is_composite) {
+    ierr = CeedOperatorGetNumSub(op, &num_suboperators); CeedChk(ierr);
+    ierr = CeedOperatorGetSubList(op, &sub_operators); CeedChk(ierr);
+    for (int k = 0; k < num_suboperators; ++k) {
+      ierr = CeedSingleOperatorAssemblyCountEntries(sub_operators[k],
+             &single_entries); CeedChk(ierr);
+      *num_entries += single_entries;
+    }
+  } else {
+    ierr = CeedSingleOperatorAssemblyCountEntries(op,
+           &single_entries); CeedChk(ierr);
+    *num_entries += single_entries;
+  }
+  ierr = CeedCalloc(*num_entries, rows); CeedChk(ierr);
+  ierr = CeedCalloc(*num_entries, cols); CeedChk(ierr);
+
+  // assemble nonzero locations
+  CeedInt offset = 0;
+  if (is_composite) {
+    ierr = CeedOperatorGetNumSub(op, &num_suboperators); CeedChk(ierr);
+    ierr = CeedOperatorGetSubList(op, &sub_operators); CeedChk(ierr);
+    for (int k = 0; k < num_suboperators; ++k) {
+      ierr = CeedSingleOperatorAssembleSymbolic(sub_operators[k], offset, *rows,
+             *cols); CeedChk(ierr);
+      ierr = CeedSingleOperatorAssemblyCountEntries(sub_operators[k],
+             &single_entries);
+      CeedChk(ierr);
+      offset += single_entries;
+    }
+  } else {
+    ierr = CeedSingleOperatorAssembleSymbolic(op, offset, *rows, *cols);
+    CeedChk(ierr);
+  }
+
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+   @brief Fully assemble the nonzero entries of a linear operator.
+
+   Expected to be used in conjunction with CeedOperatorLinearAssembleSymbolic()
+
+   The assembly routines use coordinate format, with num_entries tuples of the
+   form (i, j, value) which indicate that value should be added to the matrix
+   in entry (i, j). Note that the (i, j) pairs are not unique and may repeat.
+   This function returns the values of the nonzero entries to be added, their
+   (i, j) locations are provided by CeedOperatorLinearAssembleSymbolic()
+
+   This will generally be slow unless your operator is low-order.
+
+   @param[in]  op      CeedOperator to assemble
+   @param[out] values  Values to assemble into matrix
+
+   @ref User
+**/
+int CeedOperatorLinearAssemble(CeedOperator op, CeedVector values) {
+  int ierr;
+  CeedInt num_suboperators, single_entries = 0;
+  CeedOperator *sub_operators;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Use backend version, if available
+  if (op->LinearAssemble) {
+    return op->LinearAssemble(op, values);
+  } else {
+    // Check for valid fallback resource
+    const char *resource, *fallback_resource;
+    ierr = CeedGetResource(op->ceed, &resource); CeedChk(ierr);
+    ierr = CeedGetOperatorFallbackResource(op->ceed, &fallback_resource);
+    CeedChk(ierr);
+    if (strcmp(fallback_resource, "") && strcmp(resource, fallback_resource)) {
+      // Fallback to reference Ceed
+      if (!op->op_fallback) {
+        ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+      }
+      // Assemble
+      return CeedOperatorLinearAssemble(op->op_fallback, values);
+    }
+  }
+
+  // if neither backend nor fallback resource provides
+  // LinearAssemble, continue with interface-level implementation
+
+  bool is_composite;
+  ierr = CeedOperatorIsComposite(op, &is_composite); CeedChk(ierr);
+
+  CeedInt offset = 0;
+  if (is_composite) {
+    ierr = CeedOperatorGetNumSub(op, &num_suboperators); CeedChk(ierr);
+    ierr = CeedOperatorGetSubList(op, &sub_operators); CeedChk(ierr);
+    for (int k = 0; k < num_suboperators; ++k) {
+      ierr = CeedSingleOperatorAssemble(sub_operators[k], offset, values);
+      CeedChk(ierr);
+      ierr = CeedSingleOperatorAssemblyCountEntries(sub_operators[k],
+             &single_entries);
+      CeedChk(ierr);
+      offset += single_entries;
+    }
+  } else {
+    ierr = CeedSingleOperatorAssemble(op, offset, values); CeedChk(ierr);
+  }
+
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Create a multigrid coarse operator and level transfer operators
+           for a CeedOperator, creating the prolongation basis from the
+           fine and coarse grid interpolation
+
+  @param[in] op_fine       Fine grid operator
+  @param[in] p_mult_fine   L-vector multiplicity in parallel gather/scatter
+  @param[in] rstr_coarse   Coarse grid restriction
+  @param[in] basis_coarse  Coarse grid active vector basis
+  @param[out] op_coarse    Coarse grid operator
+  @param[out] op_prolong   Coarse to fine operator
+  @param[out] op_restrict  Fine to coarse operator
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorMultigridLevelCreate(CeedOperator op_fine,
+                                     CeedVector p_mult_fine,
+                                     CeedElemRestriction rstr_coarse, CeedBasis basis_coarse,
+                                     CeedOperator *op_coarse, CeedOperator *op_prolong,
+                                     CeedOperator *op_restrict) {
+  int ierr;
+  Ceed ceed;
+  ierr = CeedOperatorGetCeed(op_fine, &ceed); CeedChk(ierr);
+
+  // Check for compatible quadrature spaces
+  CeedBasis basis_fine;
+  ierr = CeedOperatorGetActiveBasis(op_fine, &basis_fine); CeedChk(ierr);
+  CeedInt Q_f, Q_c;
+  ierr = CeedBasisGetNumQuadraturePoints(basis_fine, &Q_f); CeedChk(ierr);
+  ierr = CeedBasisGetNumQuadraturePoints(basis_coarse, &Q_c); CeedChk(ierr);
+  if (Q_f != Q_c)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_DIMENSION,
+                     "Bases must have compatible quadrature spaces");
+  // LCOV_EXCL_STOP
+
+  // Coarse to fine basis
+  CeedInt P_f, P_c, Q = Q_f;
+  bool is_tensor_f, is_tensor_c;
+  ierr = CeedBasisIsTensor(basis_fine, &is_tensor_f); CeedChk(ierr);
+  ierr = CeedBasisIsTensor(basis_coarse, &is_tensor_c); CeedChk(ierr);
+  CeedScalar *interp_c, *interp_f, *interp_c_to_f, *tau;
+  if (is_tensor_f && is_tensor_c) {
+    ierr = CeedBasisGetNumNodes1D(basis_fine, &P_f); CeedChk(ierr);
+    ierr = CeedBasisGetNumNodes1D(basis_coarse, &P_c); CeedChk(ierr);
+    ierr = CeedBasisGetNumQuadraturePoints1D(basis_coarse, &Q); CeedChk(ierr);
+  } else if (!is_tensor_f && !is_tensor_c) {
+    ierr = CeedBasisGetNumNodes(basis_fine, &P_f); CeedChk(ierr);
+    ierr = CeedBasisGetNumNodes(basis_coarse, &P_c); CeedChk(ierr);
+  } else {
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_MINOR,
+                     "Bases must both be tensor or non-tensor");
+    // LCOV_EXCL_STOP
+  }
+
+  ierr = CeedMalloc(Q*P_f, &interp_f); CeedChk(ierr);
+  ierr = CeedMalloc(Q*P_c, &interp_c); CeedChk(ierr);
+  ierr = CeedCalloc(P_c*P_f, &interp_c_to_f); CeedChk(ierr);
+  ierr = CeedMalloc(Q, &tau); CeedChk(ierr);
+  if (is_tensor_f) {
+    memcpy(interp_f, basis_fine->interp_1d, Q*P_f*sizeof basis_fine->interp_1d[0]);
+    memcpy(interp_c, basis_coarse->interp_1d,
+           Q*P_c*sizeof basis_coarse->interp_1d[0]);
+  } else {
+    memcpy(interp_f, basis_fine->interp, Q*P_f*sizeof basis_fine->interp[0]);
+    memcpy(interp_c, basis_coarse->interp, Q*P_c*sizeof basis_coarse->interp[0]);
+  }
+
+  // -- QR Factorization, interp_f = Q R
+  ierr = CeedQRFactorization(ceed, interp_f, tau, Q, P_f); CeedChk(ierr);
+
+  // -- Apply Qtranspose, interp_c = Qtranspose interp_c
+  ierr = CeedHouseholderApplyQ(interp_c, interp_f, tau, CEED_TRANSPOSE,
+                               Q, P_c, P_f, P_c, 1); CeedChk(ierr);
+
+  // -- Apply Rinv, interp_c_to_f = Rinv interp_c
+  for (CeedInt j=0; j<P_c; j++) { // Column j
+    interp_c_to_f[j+P_c*(P_f-1)] = interp_c[j+P_c*(P_f-1)]/interp_f[P_f*P_f-1];
+    for (CeedInt i=P_f-2; i>=0; i--) { // Row i
+      interp_c_to_f[j+P_c*i] = interp_c[j+P_c*i];
+      for (CeedInt k=i+1; k<P_f; k++)
+        interp_c_to_f[j+P_c*i] -= interp_f[k+P_f*i]*interp_c_to_f[j+P_c*k];
+      interp_c_to_f[j+P_c*i] /= interp_f[i+P_f*i];
+    }
+  }
+  ierr = CeedFree(&tau); CeedChk(ierr);
+  ierr = CeedFree(&interp_c); CeedChk(ierr);
+  ierr = CeedFree(&interp_f); CeedChk(ierr);
+
+  // Complete with interp_c_to_f versions of code
+  if (is_tensor_f) {
+    ierr = CeedOperatorMultigridLevelCreateTensorH1(op_fine, p_mult_fine,
+           rstr_coarse, basis_coarse, interp_c_to_f, op_coarse, op_prolong, op_restrict);
+    CeedChk(ierr);
+  } else {
+    ierr = CeedOperatorMultigridLevelCreateH1(op_fine, p_mult_fine,
+           rstr_coarse, basis_coarse, interp_c_to_f, op_coarse, op_prolong, op_restrict);
+    CeedChk(ierr);
+  }
+
+  // Cleanup
+  ierr = CeedFree(&interp_c_to_f); CeedChk(ierr);
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Create a multigrid coarse operator and level transfer operators
+           for a CeedOperator with a tensor basis for the active basis
+
+  @param[in] op_fine        Fine grid operator
+  @param[in] p_mult_fine    L-vector multiplicity in parallel gather/scatter
+  @param[in] rstr_coarse    Coarse grid restriction
+  @param[in] basis_coarse   Coarse grid active vector basis
+  @param[in] interp_c_to_f  Matrix for coarse to fine interpolation
+  @param[out] op_coarse     Coarse grid operator
+  @param[out] op_prolong    Coarse to fine operator
+  @param[out] op_restrict   Fine to coarse operator
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorMultigridLevelCreateTensorH1(CeedOperator op_fine,
+    CeedVector p_mult_fine, CeedElemRestriction rstr_coarse, CeedBasis basis_coarse,
+    const CeedScalar *interp_c_to_f, CeedOperator *op_coarse,
+    CeedOperator *op_prolong, CeedOperator *op_restrict) {
+  int ierr;
+  Ceed ceed;
+  ierr = CeedOperatorGetCeed(op_fine, &ceed); CeedChk(ierr);
+
+  // Check for compatible quadrature spaces
+  CeedBasis basis_fine;
+  ierr = CeedOperatorGetActiveBasis(op_fine, &basis_fine); CeedChk(ierr);
+  CeedInt Q_f, Q_c;
+  ierr = CeedBasisGetNumQuadraturePoints(basis_fine, &Q_f); CeedChk(ierr);
+  ierr = CeedBasisGetNumQuadraturePoints(basis_coarse, &Q_c); CeedChk(ierr);
+  if (Q_f != Q_c)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_DIMENSION,
+                     "Bases must have compatible quadrature spaces");
+  // LCOV_EXCL_STOP
+
+  // Coarse to fine basis
+  CeedInt dim, num_comp, num_nodes_c, P_1d_f, P_1d_c;
+  ierr = CeedBasisGetDimension(basis_fine, &dim); CeedChk(ierr);
+  ierr = CeedBasisGetNumComponents(basis_fine, &num_comp); CeedChk(ierr);
+  ierr = CeedBasisGetNumNodes1D(basis_fine, &P_1d_f); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetElementSize(rstr_coarse, &num_nodes_c);
+  CeedChk(ierr);
+  P_1d_c = dim == 1 ? num_nodes_c :
+           dim == 2 ? sqrt(num_nodes_c) :
+           cbrt(num_nodes_c);
+  CeedScalar *q_ref, *q_weight, *grad;
+  ierr = CeedCalloc(P_1d_f, &q_ref); CeedChk(ierr);
+  ierr = CeedCalloc(P_1d_f, &q_weight); CeedChk(ierr);
+  ierr = CeedCalloc(P_1d_f*P_1d_c*dim, &grad); CeedChk(ierr);
+  CeedBasis basis_c_to_f;
+  ierr = CeedBasisCreateTensorH1(ceed, dim, num_comp, P_1d_c, P_1d_f,
+                                 interp_c_to_f, grad, q_ref, q_weight, &basis_c_to_f);
+  CeedChk(ierr);
+  ierr = CeedFree(&q_ref); CeedChk(ierr);
+  ierr = CeedFree(&q_weight); CeedChk(ierr);
+  ierr = CeedFree(&grad); CeedChk(ierr);
+
+  // Core code
+  ierr = CeedOperatorMultigridLevel_Core(op_fine, p_mult_fine, rstr_coarse,
+                                         basis_coarse, basis_c_to_f, op_coarse,
+                                         op_prolong, op_restrict);
+  CeedChk(ierr);
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Create a multigrid coarse operator and level transfer operators
+           for a CeedOperator with a non-tensor basis for the active vector
+
+  @param[in] op_fine        Fine grid operator
+  @param[in] p_mult_fine    L-vector multiplicity in parallel gather/scatter
+  @param[in] rstr_coarse    Coarse grid restriction
+  @param[in] basis_coarse   Coarse grid active vector basis
+  @param[in] interp_c_to_f  Matrix for coarse to fine interpolation
+  @param[out] op_coarse     Coarse grid operator
+  @param[out] op_prolong    Coarse to fine operator
+  @param[out] op_restrict   Fine to coarse operator
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref User
+**/
+int CeedOperatorMultigridLevelCreateH1(CeedOperator op_fine,
+                                       CeedVector p_mult_fine,
+                                       CeedElemRestriction rstr_coarse,
+                                       CeedBasis basis_coarse,
+                                       const CeedScalar *interp_c_to_f,
+                                       CeedOperator *op_coarse,
+                                       CeedOperator *op_prolong,
+                                       CeedOperator *op_restrict) {
+  int ierr;
+  Ceed ceed;
+  ierr = CeedOperatorGetCeed(op_fine, &ceed); CeedChk(ierr);
+
+  // Check for compatible quadrature spaces
+  CeedBasis basis_fine;
+  ierr = CeedOperatorGetActiveBasis(op_fine, &basis_fine); CeedChk(ierr);
+  CeedInt Q_f, Q_c;
+  ierr = CeedBasisGetNumQuadraturePoints(basis_fine, &Q_f); CeedChk(ierr);
+  ierr = CeedBasisGetNumQuadraturePoints(basis_coarse, &Q_c); CeedChk(ierr);
+  if (Q_f != Q_c)
+    // LCOV_EXCL_START
+    return CeedError(ceed, CEED_ERROR_DIMENSION,
+                     "Bases must have compatible quadrature spaces");
+  // LCOV_EXCL_STOP
+
+  // Coarse to fine basis
+  CeedElemTopology topo;
+  ierr = CeedBasisGetTopology(basis_fine, &topo); CeedChk(ierr);
+  CeedInt dim, num_comp, num_nodes_c, num_nodes_f;
+  ierr = CeedBasisGetDimension(basis_fine, &dim); CeedChk(ierr);
+  ierr = CeedBasisGetNumComponents(basis_fine, &num_comp); CeedChk(ierr);
+  ierr = CeedBasisGetNumNodes(basis_fine, &num_nodes_f); CeedChk(ierr);
+  ierr = CeedElemRestrictionGetElementSize(rstr_coarse, &num_nodes_c);
+  CeedChk(ierr);
+  CeedScalar *q_ref, *q_weight, *grad;
+  ierr = CeedCalloc(num_nodes_f, &q_ref); CeedChk(ierr);
+  ierr = CeedCalloc(num_nodes_f, &q_weight); CeedChk(ierr);
+  ierr = CeedCalloc(num_nodes_f*num_nodes_c*dim, &grad); CeedChk(ierr);
+  CeedBasis basis_c_to_f;
+  ierr = CeedBasisCreateH1(ceed, topo, num_comp, num_nodes_c, num_nodes_f,
+                           interp_c_to_f, grad, q_ref, q_weight, &basis_c_to_f);
+  CeedChk(ierr);
+  ierr = CeedFree(&q_ref); CeedChk(ierr);
+  ierr = CeedFree(&q_weight); CeedChk(ierr);
+  ierr = CeedFree(&grad); CeedChk(ierr);
+
+  // Core code
+  ierr = CeedOperatorMultigridLevel_Core(op_fine, p_mult_fine, rstr_coarse,
+                                         basis_coarse, basis_c_to_f, op_coarse,
+                                         op_prolong, op_restrict);
+  CeedChk(ierr);
+  return CEED_ERROR_SUCCESS;
+}
+
+/**
+  @brief Build a FDM based approximate inverse for each element for a
+           CeedOperator
+
+  This returns a CeedOperator and CeedVector to apply a Fast Diagonalization
+    Method based approximate inverse. This function obtains the simultaneous
+    diagonalization for the 1D mass and Laplacian operators,
+      M = V^T V, K = V^T S V.
+    The assembled QFunction is used to modify the eigenvalues from simultaneous
+    diagonalization and obtain an approximate inverse of the form
+      V^T S^hat V. The CeedOperator must be linear and non-composite. The
+    associated CeedQFunction must therefore also be linear.
+
+  @param op            CeedOperator to create element inverses
+  @param[out] fdm_inv  CeedOperator to apply the action of a FDM based inverse
+                         for each element
+  @param request       Address of CeedRequest for non-blocking completion, else
+                         @ref CEED_REQUEST_IMMEDIATE
+
+  @return An error code: 0 - success, otherwise - failure
+
+  @ref Backend
+**/
+int CeedOperatorCreateFDMElementInverse(CeedOperator op, CeedOperator *fdm_inv,
+                                        CeedRequest *request) {
+  int ierr;
+  ierr = CeedOperatorCheckReady(op); CeedChk(ierr);
+
+  // Use backend version, if available
+  if (op->CreateFDMElementInverse) {
+    ierr = op->CreateFDMElementInverse(op, fdm_inv, request); CeedChk(ierr);
+  } else {
+    // Fallback to reference Ceed
+    if (!op->op_fallback) {
+      ierr = CeedOperatorCreateFallback(op); CeedChk(ierr);
+    }
+    // Assemble
+    ierr = op->op_fallback->CreateFDMElementInverse(op->op_fallback, fdm_inv,
+           request); CeedChk(ierr);
+  }
+  return CEED_ERROR_SUCCESS;
+}
+
+/// @}
